fixed crash

+++ code/User_UI/user_ui.cpp
@@ -310,22 +310,20 @@ inline void showSEEK(SETTINGS_UI_MENU_LIST element)  // Prikaze element v seeku 
         break;
 
         case SU_STATE_STRIP_SELECTION:
-            if (element.index == SU_MENU_STRIP_ANIMATION_STRIP_OFF)
+            if (element.index == SU_MENU_STRIP_ANIMATION_STRIP_OFF && m_audio_system.strip.curr_color_index != COLOR_RED)
             {
-                if (m_audio_system.strip.curr_color_index != COLOR_RED)
-                {
-                    deleteTASK(&m_audio_system.handle_active_strip_mode);
-                    brightDOWN(AUVS_CFG_SLOW_ANIMATION_TIME_MS);
-                    m_audio_system.set_strip_brightness(255);
-                    m_audio_system.set_strip_color(COLOR_RED);
-                    brightUP(AUVS_CFG_SLOW_ANIMATION_TIME_MS);
-                }
+                deleteTASK(&m_audio_system.handle_active_strip_mode);
+                brightDOWN(AUVS_CFG_SLOW_ANIMATION_TIME_MS);
+                m_audio_system.set_strip_brightness(255);
+                m_audio_system.set_strip_color(COLOR_RED);
+                brightUP(AUVS_CFG_SLOW_ANIMATION_TIME_MS);
             }
 
-            else if (m_audio_system.handle_active_strip_mode == NULL)
+            else if (m_audio_system.handle_active_strip_mode == NULL && element.index != SU_MENU_STRIP_ANIMATION_STRIP_OFF)
             {
                 m_audio_system.create_animation(element.index, COLOR_WHITE);
             }
+
         break;
 	}
 }
